requirements.txt
cv_bridge==1.13.0.post0
gradio==5.23.3
numpy==2.2.4
opencv_python==4.10.0.84
Pillow==11.1.0
torch==2.5.1+cu124
transformers==4.47.1
transformers==4.45.2
ultralytics==8.3.28